requirements.txt
flask
twilio
openai
python-dotenv
streamlit
pandas
numpy